requirements.txt
Flask==2.0.1
PyPDF2==1.26.0
cloudinary==1.33.0
python-dotenv==0.19.0
gunicorn==20.1.0
img2pdf==0.4.4